% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logiSieve.R
\name{logiSieve}
\alias{logiSieve}
\title{Sieve maximum likelihood estimator (SMLE) for two-phase logistic regression problems with covariate measurement error}
\usage{
logiSieve(
  analysis_formula,
  error_formula,
  data,
  analysis_link = "logit",
  initial_lr_params = "Zero",
  pert_scale = 1,
  no_se = FALSE,
  tol = 1e-04,
  max_iter = 1000,
  output = "logORs"
)
}
\arguments{
\item{analysis_formula}{formula, analysis model formula (or coercible to formula), a formula expression as for other regression models. The response should be the logistic regression model outcome.}

\item{error_formula}{formula, covariate error model formula (or coercible to formula), a formula expression as for other regression models. The response should be the error-free version of the error-prone of the covariate, and the covariate should be the names of the B-spline columns.}

\item{data}{dataframe, a dataframe with one row per subject containing all variables from \code{analysis_formula} and \code{error_formula}.}

\item{analysis_link}{string, for logistic regression analysis model \code{link = "logit"} (the default) and for log-binomial regression \code{link = "log"}.}

\item{initial_lr_params}{character, initial values for parametric model parameters. Choices include (1) \code{"Zero"} (non-informative starting values) or (2) \code{"Complete-data"} (estimated based on validated subjects only)}

\item{pert_scale}{scalar, size of the perturbation used in estimating the standard errors via profile likelihood. If none is supplied, default is \code{pert_scale = 1}.}

\item{no_se}{logical, indicator for whether standard errors are desired. Defaults to \code{no_se = FALSE}.}

\item{tol}{scalar, tolerance between iterations in the EM algorithm used to define convergence.}

\item{max_iter}{scalar, maximum number of iterations allowed in the EM algorithm.}

\item{output}{character, level of fitted model output to be returned. Defaults to \code{output = "logORs"}, but \code{output = "all"} is also possible.}
}
\value{
A list with the following named slots:
\item{model_coeff}{dataframe with final model coefficients and standard error estimates (where applicable) for the analysis model.}
\item{bspline_coeff}{dataframe with B-spline coefficients for the covariate error model. (Only returned if \code{output = "all"}.)}
\item{vcov}{covariance matrix of \code{model_coeff} for the analysis model.}
\item{predicted}{vector with predictions for the error-free versions of covariates for unvalidated subjects. For validated subjects, their validated covariate is returned. (Only returned if \code{output = "all"}.)}
\item{converged}{indicator of EM algorithm convergence for parameter estimates.}
\item{se_converged}{indicator of standard error estimate convergence.}
\item{converged_msg}{(where applicable) description of non-convergence.}
\item{iterations}{number of iterations completed by EM algorithm to find parameter estimates. (Only returned if \code{output = "all"}.)}
\item{od_loglik_at_conv}{value of the observed-data log-likelihood at convergence. (Only returned if \code{output = "all"}.)}
}
\description{
This function returns the sieve maximum likelihood estimators (SMLE) for the logistic regression model from Lotspeich et al. (2025+).
}
