% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_data_loglik.R
\name{observed_data_loglik}
\alias{observed_data_loglik}
\title{Observed-data log-likelihood}
\usage{
observed_data_loglik(
  N,
  n,
  Y = NULL,
  X_val = NULL,
  C = NULL,
  Bspline = NULL,
  comp_dat_all,
  theta,
  p
)
}
\arguments{
\item{N}{Phase I sample size}

\item{n}{Phase II sample size}

\item{Y}{Column names with the validated outcome.}

\item{X_val}{Column name(s) with the validated predictors.}

\item{C}{(Optional) Column name(s) with additional error-free covariates.}

\item{Bspline}{Vector of column names containing the B-spline basis functions.}

\item{comp_dat_all}{Augmented dataset containing rows for each combination of unvalidated subjects' data with values from Phase II (a matrix)}

\item{theta}{Parameters for the analysis model (a column vector)}

\item{p}{B-spline coefficients for the approximated covariate error model (a matrix)}
}
\value{
Scalar value of the function
}
\description{
This function returns the value of the observed-data log-likelihood (equation (2) in Lotspeich et al. (2021))
for a given dataset and parameter values \code{theta} and \code{p}.
}
