% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_theta.R
\name{pl_theta}
\alias{pl_theta}
\title{Profile likelihood for theta, the analysis model parameters}
\usage{
pl_theta(
  k,
  theta,
  h_N,
  N,
  n,
  Y,
  X_val,
  C,
  Bspline,
  comp_dat_all,
  p0 = NULL,
  p_val_num = NULL,
  tol,
  max_iter,
  analysis_link
)
}
\arguments{
\item{k}{A numeric index between 1 and the dimension of theta for the element of theta to be perturbed}

\item{theta}{Parameters for the analysis model (a column vector) at convergence, resulting from the EM algorithm}

\item{h_N}{Size of the small perturbation in \code{theta[k]}, by default chosen to be \code{h_N =  N ^ ( - 1 / 2)}}

\item{N}{Phase I sample size}

\item{n}{Phase II sample size}

\item{Y}{Column names with the validated outcome.}

\item{X_val}{Column name(s) with the validated predictors.}

\item{C}{(Optional) Column name(s) with additional error-free covariates.}

\item{Bspline}{Vector of column names containing the B-spline basis functions.}

\item{comp_dat_all}{Augmented dataset containing rows for validated subjects and each combination of unvalidated subjects' data with values from Phase II}

\item{p0}{Starting values for \code{p}, the B-spline coefficients for the approximated covariate error model (a matrix)}

\item{p_val_num}{Contributions of validated subjects to the numerator for \code{p}, which are fixed (a matrix)}

\item{tol}{tolerance between iterations in the EM algorithm used to define convergence.}

\item{max_iter}{Maximum number of iterations allowed in the EM algorithm.}

\item{analysis_link}{String specifying link function for the analysis model}
}
\value{
Profile likelihood for \code{theta} after perturbing element \code{k} by \code{h_N}.
}
\description{
This function returns the value of the profile log-likelihood for parameters \code{theta} of the
analysis model P(Y|X,C) after perturbing element \code{k} of \code{theta} by some small amount \code{h_N}.
}
