% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ali_data.R
\name{sim_ali_data}
\alias{sim_ali_data}
\title{Generate error-prone, partially validated allostatic load index (ALI) and healthcare utilization data}
\usage{
sim_ali_data(
  tpr = 0.9,
  fpr = 0.1,
  audit_recovery = 1,
  N = 1000,
  n = 100,
  lambda_age = 4.566,
  betas = c(-1.93, 1.88, 0.1),
  pS = c(0.25, 0.987013, 0.4549098, 0.145, 0.058, 0.2490119, 0.3138501, 0.3316391,
    0.3111111, 0),
  pM = c(0.996, 0.153, 0.002, 0, 0, 0.494, 0.213, 0.213, 0.955, 0.983)
)
}
\arguments{
\item{tpr}{true positive rate for the error-prone ALI. Default is \code{tpr = 0.9}.}

\item{fpr}{false positive rate for the error-prone ALI. Default is \code{fpr = 0.1}.}

\item{audit_recovery}{proportion of missing data recovered through the validation study. Default is \code{audit_recovery = 1}.}

\item{N}{total sample size (phase I) for the error-prone EHR data. Default is \code{N = 1000}.}

\item{n}{validation sample size (phase II) for the chart review data (must have \code{N > n}). Default is \code{n = 100}.}

\item{lambda_age}{mean of the Poisson distribution for age. Default is \code{lambda_age = 4.566}.}

\item{betas}{vector of 3 coefficients for the analysis model (intercept, log odds ratio for X, and log odds ratio for Z). Default is \code{betas = c(-1.93, 1.88, 0.10)}.}

\item{pS}{vector of 10 probabilities for the components of the ALI being "yes". Default is \code{pS = c(0.2500000, 0.9870130, 0.4549098, 0.1450000, 0.0580000, 0.2490119, 0.3138501, 0.3316391, 0.3111111, 0.0000000)}.}

\item{pM}{vector of 10 probabilities for the components of the ALI being missing. Default is \code{pM = c(0.996, 0.153, 0.002, 0.000, 0.000, 0.494, 0.213, 0.213, 0.955, 0.983)}.}
}
\value{
dataframe with the following columns:
\item{id}{Simulated identifiers}
\item{X}{True ALI}
\item{Xstar}{Unvalidated ALI (from the EHR)}
\item{Xval}{Validated ALI (from the chart review)}
\item{Y}{Indicator of healthcare utilization}
\item{Z}{Age at first encounter}
}
\description{
Generate error-prone, partially validated allostatic load index (ALI) and healthcare utilization data
}
