% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_out.R
\name{profile_out}
\alias{profile_out}
\title{Profiles out nuisance parameters from the observed-data log-likelihood for a given value of theta}
\usage{
profile_out(
  theta,
  n,
  N,
  pY_X,
  Bspline = NULL,
  comp_dat_all,
  p0,
  p_val_num,
  TOL,
  MAX_ITER
)
}
\arguments{
\item{theta}{Parameters for the analysis model (a column vector)}

\item{n}{Phase II sample size}

\item{N}{Phase I sample size}

\item{pY_X}{P(Y|X) for unvalidated rows at convergence for \code{theta}.}

\item{Bspline}{Vector of column names containing the B-spline basis functions.}

\item{comp_dat_all}{Augmented dataset containing rows for each combination of unvalidated subjects' data with values from Phase II (a matrix)}

\item{p0}{Starting values for \code{p}, the B-spline coefficients for the approximated covariate error model (a matrix)}

\item{p_val_num}{Contributions of validated subjects to the numerator for \code{p}, which are fixed (a matrix)}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{MAX_ITER}{Maximum number of iterations allowed in the EM algorithm.}
}
\value{
Profile likelihood for \code{theta}: the value of the observed-data log-likelihood after profiling out other parameters.
}
\description{
For a given vector \code{theta} to parameterize P(Y|X,C), this function repeats the EM algorithm to find
the values of \code{p} at convergence. The resulting parameters are used to find the profile
log-likelihood for \code{theta} by plugging them into the observed-data log-likelihood.
This function is used by \code{pl_theta()}.
}
